+++ tests/integration/test_script_runner.py
@@ -19,7 +19,7 @@ import pytest
 from assets.scripts.build_gallery import execute_shell_command
 from great_expectations.data_context.util import file_relative_path
 from tests.integration.backend_dependencies import BackendDependencies
-from tests.integration.test_definitions.multiple_db.integration_tests import (
+from tests.integration.test_definitions.multiple_backend.integration_tests import (
     multiple_backend,
 )
 from tests.integration.test_definitions.mysql.integration_tests import (
